Initial commit from hackathon
Trello: https://trello.com/c/KqEDQxg4

This commit is an almost straight copy from the kingdom code hackathon
work[1]. I've updated paths from the the old repo name and added the
build folder to `.gitignore` as it is a generated folder that will keep
changing.

This commit only takes the front-end assets from the work we did at
kingdom code build, with the aim of splitting FE and BE into separate
repos.

1. https://github.com/andymorton/SafeCheck